refactor client to use jsx
