+++ client/package.json
@@ -9,20 +9,15 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^29.5.3",
-    "@types/node": "^20.4.1",
-    "@types/react": "18.2.1",
-    "@types/react-dom": "18.2.1",
+    "daisyui": "^3.5.0",
     "graphql": "^15.4.0",
     "jwt-decode": "^3.1.2",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.1",
-    "tailwindcss": "^3.3.2",
-    "typescript": "^5.1.6",
-    "web-vitals": "^0.2.4",
-    "daisyui": "^3.5.0"
+    "tailwindcss": "^3.3.3",
+    "web-vitals": "^0.2.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -47,9 +42,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@types/moment": "^2.13.0",
-    "daisyui": "^3.5.0"
   }
 }
