+++ client/src/components/common/Stripe/index.jsx
@@ -10,7 +10,7 @@ const ProductDisplay = () => (
   </section>
 );
 
-const Message = ({ message }: any) => (
+const Message = ({ message }) => (
 <section>
     <p className="break">{message}</p>
 </section>
@@ -20,21 +20,21 @@ export default function Stripe () {
     const [message, setMessage] = useState("");
 
     useEffect(() => {
-      // Check to see if this is a redirect back from Checkout
-      const query = new URLSearchParams(window.location.search);
-  
-      if (query.get("success")) {
+        // Check to see if this is a redirect back from Checkout
+        const query = new URLSearchParams(window.location.search);
+
+        if (query.get("success")) {
         setMessage("Thank you for donating to my alcohol fund.");
-      }
-      
+        }
+        
     }, []);
-  
+
     return (
-      <>
+        <>
         <ProductDisplay />
         {message ? (
-          <Message message={message} />
+            <Message message={message} />
         ) : null }
-      </>
+        </>
     );
-  }
+}
